+++ ultratrader-go/internal/exchange/binance/adapter.go
@@ -57,6 +57,8 @@ func (a *Adapter) Name() string { return "binance" }
 
 func (a *Adapter) SetBaseURL(url string) { a.baseURL = url }
 
+func (a *Adapter) IsTestnet() bool { return a.config.Testnet }
+
 func (a *Adapter) Capabilities() []exchange.Capability {
 	return []exchange.Capability{
 		exchange.CapabilitySpot,